git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@2314 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Main/Base/Project/Src/Gui/Components/FontSelectionPanel.cs
@@ -199,9 +199,9 @@ namespace ICSharpCode.SharpDevelop.Gui
 		{
 			if (!fontListComboBox.Enabled)
 				return null;
-			int fontSize = 10;
+			float fontSize = 10f;
 			try {
-				fontSize = Math.Max(6, Int32.Parse(fontSizeComboBox.Text));
+				fontSize = Math.Max(6, Single.Parse(fontSizeComboBox.Text));
 			} catch (Exception) {}
 			
 			FontDescriptor fontDescriptor = (FontDescriptor)fontListComboBox.Items[fontListComboBox.SelectedIndex];
